server/requirements.txt
Flask~=3.1.0
flask_session~=0.8.0
playwright~=1.50.0
bs4~=0.0.2
beautifulsoup4~=4.13.3
lxml
cloudscraper~=1.2.71
mysql-connector-python
scikit-learn
sentence-transformers
pandas
numpy
spacy
evaluate
datasets